% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_06Q.R
\name{cov06Q}
\alias{cov06Q}
\title{Covariance Estimation via Nearest Positive-Definite Matrix Projection}
\usage{
cov06Q(X)
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix whose rows are observations.}
}
\value{
a named list containing: \describe{
\item{S}{a \eqn{(p\times p)} covariance matrix.}
}
}
\description{
Qi and Sun (2006) proposed an algorithm for computing the positive correlation matrix
with \emph{Positive Definiteness} and transforming it back in order to estimate covariance matrix.
}
\examples{
## generate a toy data from standard normal
n   = 10
p   = 20
dat = matrix(stats::rnorm(n*p), ncol=p)

## compare against sample covariance & true model
out1 <- coper::cov(dat)$S
out2 <- coper::cov06Q(dat)$S
out3 <- diag(p)

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,3), pty="s")
image(out1, xaxt='n', yaxt='n', main="sample covariance")
image(out2, xaxt='n', yaxt='n', main="06Q estimate")
image(out3, xaxt='n', yaxt='n', main="model")
par(opar)

}
\references{
\insertRef{qi_quadratically_2006}{coper}
}
\concept{covariance}
