% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_cov2corr.R
\name{cov2corr}
\alias{cov2corr}
\title{Convert Covariance to Correlation}
\usage{
cov2corr(Sigma)
}
\arguments{
\item{Sigma}{an \eqn{(p\times p)} symmetric matrix of positive diagonal entries.}
}
\value{
a \eqn{(p\times p)} symmetric matrix of unit diagonals
}
\description{
Given a symmetric matrix of positive diagonal entries, this function scales the 
matrix so that it has unit diagonals.
}
\examples{
\donttest{
## generate a toy data from 5-dimensional standard normal
n   = 25
p   = 5
dat = matrix(stats::rnorm(n*p), ncol=p)

## compute sample covariance
S = coper::cov(dat)

## convert into correlation
C = coper::cov2corr(S)

## visualize
opar <- par(no.readonly=TRUE) 
par(mfrow=c(1,2), pty="s")
image(S, main="covariance")
image(C, main="correlation")
par(opar)
}

}
\concept{utility}
