% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_cov2pcor.R
\name{cov2pcor}
\alias{cov2pcor}
\title{Partial correlation given a covariance matrix}
\usage{
cov2pcor(Sigma)
}
\description{
If a covariance \eqn{\Sigma} is given, a partial correlation matrix is computed. 
Note that this function strictly requires a given input to be symmetric and 
positive-definite since it relies on matrix inversion.
}
\concept{utility}
