% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_pcor.R
\name{pcor}
\alias{pcor}
\title{Partial Correlation}
\usage{
pcor(X)
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix whose rows are observations.}
}
\value{
a \eqn{(p\times p)} partial correlation matrix.
}
\description{
Compute a partial correlation matrix.
}
\examples{
\donttest{
## generate a toy data from 5-dimensional standard normal
n   = 25
p   = 5
dat = matrix(stats::rnorm(n*p), ncol=p)

## compute partial correlation
C = coper::corr(dat)

## visualize
opar <- par(no.readonly=TRUE)
par(pty="s")
image(C, main="partial correlation")
par(opar)
}

}
\concept{correlation}
