% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_cov.R
\name{cov}
\alias{cov}
\title{(Sample) Covariance}
\usage{
cov(X)
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix whose rows are observations.}
}
\value{
a \eqn{(p\times p)} covariance matrix.
}
\description{
Compute an empirical sample covariance.
}
\examples{
\donttest{
## generate a toy data from 5-dimensional standard normal
n   = 25
p   = 5
dat = matrix(stats::rnorm(n*p), ncol=p)

## compute sample covariance
S = coper::cov(dat)

## visualize
opar <- par(no.readonly=TRUE)
par(pty="s")
image(S, main="sample covariance")
par(opar)
}

}
\concept{covariance}
