% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_pcor.R
\name{pcor}
\alias{pcor}
\title{Partial Correlation}
\usage{
pcor(X)
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix whose rows are observations.}
}
\value{
a named list containing: \describe{
\item{C}{a \eqn{(p\times p)} partial correlation matrix.}
}
}
\description{
Compute a partial correlation matrix using the regression method.
}
\examples{
## generate a toy data from 5-dimensional standard normal
n   = 25
p   = 5
dat = matrix(stats::rnorm(n*p), ncol=p)

## compute partial correlation
C = coper::corr(dat)$C

## true model parameter from standard normal
I = diag(5)

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
image(I, xaxt='n', yaxt='n', main="model")
image(C, xaxt='n', yaxt='n', main="partial correlation")
par(opar)

}
\concept{correlation}
